requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-dev --no-hashes --output-file requirements.txt
asgiref==3.11.1
    # via django
django==6.0.4
    # via
    #   django-jazzmin
    #   django-ranged-response
    #   django-recaptcha
    #   django-simple-captcha
    #   forum
django-jazzmin==3.0.4
    # via forum
django-ranged-response==0.2.0
    # via django-simple-captcha
django-recaptcha==4.1.0
    # via forum
django-simple-captcha==0.6.3
    # via forum
django-uuslug==2.0.0
    # via forum
gunicorn==25.3.0
    # via forum
packaging==26.1
    # via gunicorn
pillow==12.2.0
    # via
    #   django-simple-captcha
    #   forum
python-dotenv==1.2.2
    # via forum
python-slugify==8.0.4
    # via django-uuslug
sqlparse==0.5.5
    # via django
text-unidecode==1.3
    # via python-slugify
tzdata==2026.1 ; sys_platform == 'win32'
    # via django
